Add files via upload
To accomplish the objectives of this project, I employed a powerful tool called Beautiful Soup - a Python library that facilitates the extraction of data from HTML files. By utilizing the requests method, I was able to pull out relevant information from these files. However, the extraction process was not straightforward - I had to thoroughly analyze the HTML page to identify the specific data that I needed. 

To further simplify the process, I created custom functions that enabled me to retrieve the data efficiently. These functions were designed to extract the desired data and organize it in a structured manner. 

Once I had obtained the data, I converted it into a data frame to allow for easy manipulation and analysis. This conversion facilitated the process of viewing the data and identifying patterns that I could further explore. Finally, I saved the data to my local computer in an Excel file for future reference and analysis.